Move silent code
