+++ lib/msal-browser/src/error/BrowserAuthError.ts
@@ -89,6 +89,10 @@ export const BrowserAuthErrorMessage = {
         code: "silent_sso_error",
         desc: "Silent SSO could not be completed - insufficient information was provided. Please provide either a loginHint or sid."
     },
+    silentLogoutUnsupportedError: {
+        code: "silent_logout_unsupported",
+        desc: "Silent logout not supported. Please call logoutRedirect or logoutPopup instead."
+    },
     noAccountError: {
         code: "no_account_error",
         desc: "No account object provided to acquireTokenSilent and no active account has been set. Please call setActiveAccount or provide an account on the request."
@@ -312,6 +316,13 @@ export class BrowserAuthError extends AuthError {
      */
     static createSilentSSOInsufficientInfoError(): BrowserAuthError {
         return new BrowserAuthError(BrowserAuthErrorMessage.silentSSOInsufficientInfoError.code, BrowserAuthErrorMessage.silentSSOInsufficientInfoError.desc);
+    }
+
+    /**
+     * Creates an error thrown when the logout API is called on any of the silent interaction clients
+     */
+    static createSilentLogoutUnsupportedError(): BrowserAuthError {
+        return new BrowserAuthError(BrowserAuthErrorMessage.silentLogoutUnsupportedError.code, BrowserAuthErrorMessage.silentLogoutUnsupportedError.desc);
     }
 
     /**
