+++ lib/msal-browser/src/interaction_client/SilentCacheClient.ts
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { StandardInteractionClient } from "./StandardInteractionClient";
+import { CommonSilentFlowRequest, AuthenticationResult, SilentFlowClient, ServerTelemetryManager, AccountInfo } from "@azure/msal-common";
+import { version } from "../packageMetadata";
+import { SilentRequest } from "../request/SilentRequest";
+import { EventType } from "../event/EventType";
+import { InteractionType, ApiId } from "../utils/BrowserConstants";
+import { BrowserAuthError } from "../error/BrowserAuthError";
+
+export class SilentCacheClient extends StandardInteractionClient {
+    /**
+     * Returns unexpired tokens from the cache, if available
+     * @param silentRequest 
+     */
+    async acquireToken(silentRequest: CommonSilentFlowRequest): Promise<AuthenticationResult> {
+        this.logger = this.logger.clone(name, version, silentRequest.correlationId);
+
+        // Telemetry manager only used to increment cacheHits here
+        const serverTelemetryManager = this.initializeServerTelemetryManager(ApiId.acquireTokenSilent_silentFlow, silentRequest.correlationId);
+        const silentAuthClient = await this.createSilentFlowClient(serverTelemetryManager, silentRequest.authority, silentRequest.correlationId);
+        this.logger.verbose("Silent auth client created");
+        const cachedToken = await silentAuthClient.acquireCachedToken(silentRequest);
+        this.eventHandler.emitEvent(EventType.ACQUIRE_TOKEN_SUCCESS, InteractionType.Silent, cachedToken);
+        return cachedToken;
+    }
+    
+    /**
+     * Currently Unsupported
+     */
+    logout(): Promise<void> {
+        throw BrowserAuthError.createSilentLogoutUnsupportedError();
+    }
+
+    /**
+     * Creates an Silent Flow Client with the given authority, or the default authority.
+     * @param serverTelemetryManager
+     * @param authorityUrl
+     */
+    protected async createSilentFlowClient(serverTelemetryManager: ServerTelemetryManager, authorityUrl?: string, correlationId?: string): Promise<SilentFlowClient> {
+        // Create auth module.
+        const clientConfig = await this.getClientConfiguration(serverTelemetryManager, authorityUrl, correlationId);
+        return new SilentFlowClient(clientConfig);
+    }
+
+    initializeSilentRequest(request: SilentRequest, account: AccountInfo): CommonSilentFlowRequest {
+        return {
+            ...request,
+            ...this.initializeBaseRequest(request),
+            account: account,
+            forceRefresh: request.forceRefresh || false
+        };
+    }
+}
